solving some lint rules errors

+++ lib/models/movie.dart
@@ -41,7 +41,7 @@ class Movie {
     originalLanguage = json['original_language'];
     title = json['title'];
     backdropPath = json['backdrop_path'];
-    popularity = json['popularity'];
+    popularity = json['popularity'].toDouble();
     voteCount = json['vote_count'];
     video = json['video'];
     voteAverage = json['vote_average'].toDouble();
